+++ addons/scene_manager/item_setting.gd
@@ -1,4 +1,3 @@
-extends Object
 class_name ItemSetting
 
 var visibility: bool = true