+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat: added ability to load scenes interactively + some more useful functions added
 
 3.2.1 (2023-01-03)
 ------------------